minor - bug - DetachedInstanceError on question log (refactor log clearing)

+++ app/updater.py
@@ -91,9 +91,13 @@ def update_TestQuestionLogs(app):
                 db.session.add(addTest)
                 db.session.commit()
                 
-                #Clear old logs if needed
-                clear_old_logs(app)
-                
+                #Delete old questions first to avoid orphaned questions (delete 1000 more than the max to avoid frequent clears)
+                cutoff = max(db.session.query(QuestionLog.id).count() + 1000 - current_app.config['MAX_QUESTIONS_LOGGED'], 0)
+                if cutoff > 0:
+                    cutoff_id = db.session.query(QuestionLog).order_by(QuestionLog.id)[cutoff].id
+                    db.session.query(QuestionLog).filter(QuestionLog.id < cutoff_id).delete()
+                    print(str(cutoff) + " old questions deleted")
+        
                 #Bulk dump the question log (only save MAX_QUESTIONS_LOGGED_EACH per test)
                 db.engine.execute(
                         QuestionLog.__table__.insert(),
@@ -238,14 +242,16 @@ def clear_old_logs(app):
         print("--------LOG CLEANUP------------")
         #Delete old questions first to avoid orphaned questions (delete 1000 more than the max to avoid frequent clears)
         cutoff = max(db.session.query(QuestionLog.id).count() + 1000 - current_app.config['MAX_QUESTIONS_LOGGED'], 0)
-        cutoff_id = db.session.query(QuestionLog).order_by(QuestionLog.id)[cutoff].id
-        db.session.query(QuestionLog).filter(QuestionLog.id < cutoff_id).delete()
-        print(str(cutoff) + " old questions deleted")
+        if cutoff > 0:
+            cutoff_id = db.session.query(QuestionLog).order_by(QuestionLog.id)[cutoff].id
+            db.session.query(QuestionLog).filter(QuestionLog.id < cutoff_id).delete()
+            print(str(cutoff) + " old questions deleted")
         
         cutoff = max(db.session.query(TestLog.id).count() + 1000 - current_app.config['MAX_TESTS_LOGGED'], 0)
-        cutoff_id = db.session.query(TestLog).order_by(TestLog.id)[cutoff].id
-        db.session.query(TestLog).filter(TestLog.id < cutoff_id).delete()
-        print(str(cutoff) + " old tests deleted")
+        if cutoff > 0:
+            cutoff_id = db.session.query(TestLog).order_by(TestLog.id)[cutoff].id
+            db.session.query(TestLog).filter(TestLog.id < cutoff_id).delete()
+            print(str(cutoff) + " old tests deleted")
         
         db.session.commit()
